pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "system-sl"
version = "0.1.0"
description = "A CLI-based Solo Leveling-inspired task management system."
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
  "pywebview",
  "PyQt5",
  "PyQtWebEngine",
  "google-api-python-client",
  "google-auth-httplib2",
  "google-auth-oauthlib",
]

[project.scripts]
system-sl = "utils.notifications:main"
system-cli = "cli.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.html", "*.css", "*.mp3"]
